requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.2.0
PyYAML==5.1.2
selenium==3.141.0
urllib3==1.25.6
Werkzeug==0.16.0
